% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation Functions.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Multivariate Normal Density function}
\usage{
dmvnorm(x, mu, sigma, logged = FALSE)
}
\description{
Calculates the (log) multivariate normal density function for a matrix of
observations. Based on the Rfast version except it does a Cholesky
decomposition on the covariance sigma, and re-uses it to compute the
log-determinant instead of starting from scratch. Will throw an error if the
Cholesky decomposition fails.
}
