% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation Functions.R
\name{squarem.wrapper}
\alias{squarem.wrapper}
\title{Wrapper for SQUAREM package function}
\usage{
squarem.wrapper(
  data,
  par,
  e.step,
  m.step,
  loglik,
  wrapper,
  unwrapper,
  pre.em = 100,
  ...
)
}
\arguments{
\item{data}{As in accelerated.em().}

\item{par}{As in accelerated.em().}

\item{e.step}{As in accelerated.em().}

\item{m.step}{As in accelerated.em().}

\item{loglik}{As in accelerated.em().}

\item{wrapper}{A function that converts the parameters, often of a list form,
to the vectorized representation expected by SQUAREM. This function should
try to preserve alternate metadata by assigning the original par to an
attribute of the parameter vector.}

\item{unwrapper}{The inverse of the wrapper function.}

\item{pre.em}{As in accelerated.em().}

\item{...}{Further arguments to be passed to SQUAREM control.}
}
\description{
Wrapper for SQUAREM package function
}
\details{
The wrapper and unwrapper functions permit re-use of the e.step, m.step, and
loglik functions used in the extrapolated conditional expectation
acceleration procedure. If dedicated e.step, m.step, and loglik functions
conforming to the forms expected by SQUAREM are available, set the wrapper
and unwrapper functions to be the identity.
}
